app/url_watch/requirements.txt
celery
django==2.2.*
django-celery-beat
djangorestframework
ipdb
ipython
psycopg2
requests
